% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sGlobe-package.R
\docType{package}
\name{sGlobe-package}
\alias{sGlobe}
\alias{sGlobe-package}
\title{sGlobe: Data Package with biomass production data}
\description{
Contains a dataset (named dat) with standing biomass in temperate grasslands
}
\keyword{internal}
