% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sglobe.dat-package.R
\docType{package}
\name{sglobe.dat-package}
\alias{sglobe.dat}
\alias{sglobe.dat-package}
\title{sglobe.dat: Data Package with biomass production data}
\description{
Contains a dataset with standing biomass in temperate grasslands
}
\keyword{internal}
